requirements.txt

tensorflow>=2.9.0
Pillow
numpy
